rag-backend/requirements.txt
fastapi
uvicorn
pymongo
langchain>=0.2
langchain-community
langchain-mongodb
google-genai
sentence-transformers
unstructured
pypdf
python-docx
python-pptx
pandas
openpyxl
python-dotenv
pymupdf          
python-docx     
python-pptx     
pandas           
openpyxl  
torch
numpy       
python-multipart
python-jose
